refactor: remove unused DHCP schemas and update import paths in related modules
